+++ backend/src/main/java/codezap/member/controller/MemberController.java
@@ -12,7 +12,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -22,10 +21,10 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import codezap.member.configuration.BasicAuthentication;
-import codezap.member.dto.LoginRequest;
-import codezap.member.dto.LoginResponse;
+import codezap.member.dto.request.LoginRequest;
+import codezap.member.dto.response.LoginResponse;
 import codezap.member.dto.MemberDto;
-import codezap.member.dto.SignupRequest;
+import codezap.member.dto.request.SignupRequest;
 import codezap.member.dto.request.UpdateMemberProfileRequest;
 import codezap.member.dto.response.FindMemberResponse;
 import codezap.member.dto.response.UpdateMemberProfileResponse;
