+++ backend/src/main/java/codezap/member/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package codezap.member.dto;
+package codezap.member.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -7,7 +7,7 @@ import jakarta.validation.constraints.Size;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record SignupRequest(
+public record LoginRequest(
         @Schema(description = "이메일", example = "code@zap.com")
         @Email(message = "이메일 형식이 아닙니다.")
         @NotBlank(message = "이메일이 입력되지 않았습니다.")
@@ -18,11 +18,6 @@ public record SignupRequest(
         @Pattern(regexp = "^(?=.*[a-zA-Z])(?=.*\\d).{8,}$", message = "영어와 숫자를 포함해야합니다.")
         @NotBlank(message = "비밀번호가 입력되지 않았습니다.")
         @Size(min = 8, max = 255, message = "비밀번호는 8~16자 사이로 입력해주세요.")
-        String password,
-
-        @Schema(description = "사용자명", example = "zappy")
-        @NotBlank(message = "사용자명이 입력되지 않았습니다.")
-        @Size(min = 2, max = 255, message = "사용자명은 2~255자 사이로 입력해주세요.")
-        String username
+        String password
 ) {
 }
